Move autopaste logic to a module
