+++ src-tauri/src/commands.rs
@@ -1,36 +1,18 @@
 use crate::{
-    emit_event, img::imagedata_to_image, master::ensure_max_items, state::AppState, Event,
+    emit_event, img::imagedata_to_image, master::ensure_max_items, state::AppState, Event, autopaste::{paste_from_clipboard, get_active_elements},
 };
 use mouse_position::mouse_position::Mouse;
-use std::{fs::File, io::Write, process::Command, thread};
+use std::{fs::File, io::Write, thread};
 use tauri::{GlobalShortcutManager, LogicalPosition, LogicalSize, Manager};
 
 fn on_shortcut(handle: tauri::AppHandle) {
     let state = handle.state::<AppState>();
     let mut app_state = state.0.lock().unwrap();
 
-    #[cfg(target_os = "macos")]
-    {
-        let front_app_script = r#"
-        tell application "System Events"
-            set frontApp to bundle identifier of first application process whose frontmost is true
-        end tell
+    let (active_window, active_element) = get_active_elements();
 
-        frontApp
-    "#;
-
-        let output = Command::new("osascript")
-            .arg("-e")
-            .arg(front_app_script)
-            .output()
-            .expect("failed to execute process");
-
-        let output = String::from_utf8(output.stdout).unwrap();
-
-        let output = output.trim();
-
-        app_state.last_active_window = Some(output.to_string());
-    }
+    app_state.last_active_window = active_window;
+    app_state.last_active_element = active_element;
 
     let mouse_position = get_mouse_position();
     let app_window = handle.get_window("main").unwrap();
@@ -41,30 +23,6 @@ fn on_shortcut(handle: tauri::AppHandle) {
     let y = mouse_position.1 as i32;
 
     let result = app_window.set_position(LogicalPosition::new(x, y));
-
-    #[cfg(target_os = "linux")]
-    {
-        let active_win_id_output = Command::new("xdotool")
-            .args(&["getactivewindow"])
-            .output()
-            .expect("Failed to get active window ID");
-        let active_win_id = String::from_utf8_lossy(&active_win_id_output.stdout)
-            .trim()
-            .to_string();
-
-        let active_element_name_output = Command::new("xprop")
-            .args(&["-id", &active_win_id, "_NET_WM_NAME"])
-            .output()
-            .expect("Failed to get active element name");
-        let active_element_name =
-            String::from_utf8_lossy(&active_element_name_output.stdout).to_string();
-
-        println!("Active window id: {}", active_win_id.to_string());
-        println!("Active element name: {}", active_element_name.to_string());
-
-        app_state.last_active_window = Some(active_win_id);
-        app_state.last_active_element = Some(active_element_name);
-    }
 
     if result.is_err() {
         eprintln!("Error: {}", result.err().unwrap());
@@ -113,135 +71,11 @@ pub fn recopy_at_index(index: usize, handle: tauri::AppHandle) {
     let timeout = std::time::Duration::from_millis(100);
     thread::sleep(timeout);
 
-    #[cfg(target_os = "windows")]
-    {
-        use winapi::um::winuser::{keybd_event, KEYEVENTF_EXTENDEDKEY, VK_CONTROL};
-        unsafe {
-            keybd_event(VK_CONTROL as u8, 0, 0, 0);
-            keybd_event(86, 0, KEYEVENTF_EXTENDEDKEY, 0);
-            keybd_event(
-                86,
-                0,
-                KEYEVENTF_EXTENDEDKEY | winapi::um::winuser::KEYEVENTF_KEYUP,
-                0,
-            );
-            keybd_event(VK_CONTROL as u8, 0, winapi::um::winuser::KEYEVENTF_KEYUP, 0);
-        }
-    }
+    let last_active_window = app_state.last_active_window.clone();
+    let last_active_element = app_state.last_active_element.clone();
 
-    #[cfg(target_os = "macos")]
-    {
-        let last_active_window = app_state.last_active_window.clone();
-
-        let refocusScript = format!(
-            r#"tell application "{}"
-            activate
-            delay 0.02
-            tell application "System Events" to keystroke "v" using command down
-        end tell"#,
-            last_active_window.unwrap()
-        );
-
-        let output = Command::new("osascript")
-            .arg("-e")
-            .arg(refocusScript)
-            .output()
-            .expect("failed to execute process");
-
-        let output = String::from_utf8(output.stdout).unwrap();
-
-        let output = output.trim();
-
-        println!("Output: {}", output);
-    }
-
-    #[cfg(target_os = "linux")]
-    {
-        sleep(Duration::from_millis(100));
-
-        let last_active_window_id = &(&app_state).last_active_window;
-        let last_active_element_name = &(&app_state).last_active_element;
-
-        if last_active_window_id.is_none() || last_active_element_name.is_none() {
-            return;
-        }
-
-        let last_active_window_id = last_active_window_id.as_ref().unwrap();
-        let last_active_element_name = last_active_element_name.as_ref().unwrap();
-
-        Command::new("xdotool")
-            .args(&["windowactivate", &last_active_window_id.to_string()])
-            .output()
-            .expect("Failed to activate previous window");
-        Command::new("xdotool")
-            .args(&["windowfocus", &last_active_window_id.to_string()])
-            .output()
-            .expect("Failed to focus previous window");
-
-        // Wait for the window to become active and then search for the element by name
-        loop {
-            let win_name_output = Command::new("xprop")
-                .args(&["-id", &last_active_window_id.to_string(), "_NET_WM_NAME"])
-                .output()
-                .expect("Failed to get window name");
-            let win_name = String::from_utf8_lossy(&win_name_output.stdout)
-                .trim()
-                .to_string();
-
-            if win_name.trim() == last_active_element_name.to_string().trim() {
-                break;
-            }
-
-            sleep(Duration::from_millis(5));
-        }
-
-        Command::new("xdotool")
-            .args(&[
-                "search",
-                "--name",
-                &last_active_element_name.to_string(),
-                "windowactivate",
-                "--sync",
-            ])
-            .output()
-            .expect("Failed to activate element window");
-
-        sleep(Duration::from_millis(20));
-
-        Command::new("xdotool")
-            .args(&["key", "ctrl+v"])
-            .output()
-            .expect("Failed to simulate Ctrl+V key press");
-    }
-
-    #[cfg(target_os = "macos")]
-    {
-        let last_active_window = app_state.last_active_window.clone();
-
-        println!("{:?}", (&last_active_window));
-
-        let refocus_script = format!(
-            r#"tell application id "{}"
-            activate
-            delay .02
-            tell application "System Events" to keystroke "v" using command down
-        end tell"#,
-            last_active_window.unwrap()
-        );
-
-        println!("{}", refocus_script);
-
-        let output = Command::new("osascript")
-            .arg("-e")
-            .arg(refocus_script)
-            .output()
-            .expect("failed to execute process");
-
-        let output = String::from_utf8(output.stdout).unwrap();
-
-        let output = output.trim();
-
-        println!("Output: {}", output);
+    if last_active_window.is_some() {
+        paste_from_clipboard(last_active_window.unwrap(), last_active_element);
     }
 }
